# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Night City RP Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/discord/callback:
    get:
      operationId: discordCallback
      tags: [auth]
      summary: Discord OAuth2 callback
      parameters:
        - { in: query, name: code, schema: { type: string } }
        - { in: query, name: state, schema: { type: string } }
      responses:
        "302": { description: Redirect after login }
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { description: Not logged in }

  # ---------------- CHARACTERS ----------------
...
  /admin/users/{userId}:
    parameters:
      - { in: path, name: userId, required: true, schema: { type: string } }
    get:
      operationId: adminGetUser
      tags: [admin]
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
...
  /admin/wallet/adjust:
    post:
      operationId: adminAdjustWallet
      tags: [admin]
      summary: Manual eddie adjustment (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WalletAdjustmentInput" }
      responses:
        "200":
          description: Adjusted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
...
components:
  schemas:
...
    Me:
      type: object
      required: [id, discordId, username, avatarUrl, roles, isAdmin, isFixer, isCsApprover, isRipperdoc, isStoreOwner]
      properties:
        id: { type: string }
        discordId: { type: string }
        username: { type: string }
        globalName: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        roles:
          type: array
          items: { type: string }
        isAdmin: { type: boolean }
        isFixer: { type: boolean }
        isCsApprover: { type: boolean }
        isRipperdoc: { type: boolean }
        isStoreOwner: { type: boolean }
        activeCharacterId: { type: ["integer", "null"] }
    Character:
      type: object
      required: [id, ownerId, name, kind, createdAt]
      properties:
        id: { type: integer }
        ownerId: { type: string }
        name: { type: string }
        kind: { type: string, enum: [pc, npc] }
        archetype: { type: ["string", "null"] }
        background: { type: ["string", "null"] }
        portraitUrl: { type: ["string", "null"] }
        discordChannelId: { type: ["string", "null"] }
        isActive: { type: boolean }
        approved: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    Wallet:
      type: object
      required: [characterId, balance, source]
      properties:
        characterId: { type: integer }
        balance: { type: integer }
        bank: { type: integer }
        cash: { type: integer }
        source: { type: string, enum: [unbelievaboat, local] }
        updatedAt: { type: string, format: date-time }
...
    AdminUser:
      type: object
      required: [id, discordId, username, roles, isAdmin]
      properties:
        id: { type: string }
        discordId: { type: string }
        username: { type: string }
        globalName: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        roles:
          type: array
          items: { type: string }
        isAdmin: { type: boolean }
        isFixer: { type: boolean }
        isCsApprover: { type: boolean }
        isRipperdoc: { type: boolean }
        isStoreOwner: { type: boolean }
        characterCount: { type: integer }
        lastSeenAt: { type: ["string", "null"], format: date-time }
        characters:
          type: array
          items: { $ref: "#/components/schemas/Character" }
    WalletAdjustmentInput:
      type: object
      required: [characterId, amount, reason]
      properties:
        characterId: { type: integer }
        amount: { type: integer, description: "positive or negative" }
        reason: { type: string, minLength: 1 }
...